+++ unikmer/cmd/count.go
@@ -586,7 +586,7 @@ func init() {
 
 	countCmd.Flags().IntP("scale", "D", 1, `scale/down-sample factor`)
 	countCmd.Flags().IntP("minimizer-w", "W", 0, `minimizer window size`)
-	countCmd.Flags().IntP("syncmer-s", "S", 0, `syncmer s`)
+	countCmd.Flags().IntP("syncmer-s", "S", 0, `bounded syncmer length`)
 
 	countCmd.Flags().BoolP("linear", "l", false, `output k-mers in linear order`)
 }